+++ src/db/schema.ts
@@ -1,3 +1,3 @@
-export { users } from 'src/user/user.schema';
+export { users } from '@app/modules/user/user.schema';
 export { health } from '@app/modules/health/health.schema';
 export { tours } from '@app/modules/tours/tours.schema';
