+++ src/modules/auth/auth.controller.ts
@@ -1,6 +1,6 @@
 import { Controller, Post, Req, UseGuards } from '@nestjs/common';
 import { AuthService } from './auth.service';
-import { JwtAuthGuard } from './jwt-auth.guard';
+import { JwtAuthGuard } from '../../common/guards/jwt-auth.guard';
 import { JWTPayload } from '@app/types/jwt.payload';
 
 @Controller('auth')
